Add SonarLint configs for all 3 Sonar projects; refactored `executable_worker` package and extracted code to different files per worker; replaced deprecated dependencies
