+++ ui/.sonarlint/connectedMode.json
@@ -1,0 +1,4 @@
+{
+    "sonarCloudOrganization": "kostiantyn-github",
+    "projectKey": "kostiantyn-github_photoview-ui"
+}
